refactor(test): share the live route-table walker (#14)
The Express route-table walk lives in test/auth-csrf.test.js, where it backs the
CSRF classification guard. The anonymous-rejection enumeration for issue #14
needs the same walk, and a second copy is a second chance for one of them to
quietly stop finding routes - which makes every test built on it pass
vacuously. Moved verbatim to test/helpers/route-table.js; the CSRF suite now
requires it and is otherwise untouched.
